frontend/requirements.txt
flask==2.1.1
flask-login==0.6.0
flask-wtf==1.0.1
requests==2.27.1
requests-toolbelt==0.9.1
mysql-connector-python==8.0.32
xmltodict==0.13.0